Fix(mcp-server): Allow float for yearsExperience and refine CV processing
- Modifies 'cv_job_data_schema' in 'schemas.ts' to allow 'yearsExperience' to be a float (by removing .int()). This resolves validation errors when NLU extracts fractional years of experience (e.g., 0.5).

- Refines the 'query' mapper in 'process-cv-job-data.ts' to use a limited number (up to 3) of unique job titles from the CV, leading to more focused search queries.

- Updates JSDoc comments for the 'mappers' object in 'process-cv-job-data.ts' for improved clarity and conciseness.

+++ packages/mcp-server/src/schemas.ts
@@ -16,7 +16,7 @@ export const cv_job_data_schema = z.object({
   jobTitles: z.array(z.string()).optional(),
   locations: z.array(z.string()).optional(),
   seniorityKeywords: z.array(z.string()).optional(),
-  yearsExperience: z.number().int().nonnegative().optional(),
+  yearsExperience: z.number().nonnegative().optional(),
   companyNames: z.array(z.string()).optional(),
   educationLevel: z.string().optional(),
   fullCvText: z.string().optional(),
